Updates after checking for errors
