+++ src/models/__init__.py
@@ -17,12 +17,10 @@ from .quantum_decoder import QuantumDecoder
 from .quantum_encoder import QuantumEncoder
 from .quantum_neural_network import QuantumNeuralNetworkModel
 from .quantum_transformer import Transformer
-from .hybrid_neural_network import ClassicalQuantumClassifier
 
 __all__ = [
     "QuantumDecoder",
     "QuantumEncoder",
     "QuantumNeuralNetworkModel",
     "Transformer",
-    "ClassicalQuantumClassifier"
 ]
