+++ tests/test_quantum_feed_forward.py
@@ -15,10 +15,18 @@
 
 # Test the FeedForwardBlock class
 import torch
+import sys
+import os
+
+# Add the src directory to the Python path
+script_dir = os.path.dirname(__file__)
+src_dir = os.path.abspath(os.path.join(script_dir, '..', 'src'))
+if src_dir not in sys.path:
+    sys.path.append(src_dir)
 from layers.quantum_feed_forward import QuantumFeedForward
+from layers.qnn_circuit import qnn_circuit
 
-
-def test_feed_forward_block():
+def test_feed_forward_block(num_layers, num_wires,quantum_nn,embed_len):
     # Define parameters
     embed_len = 64
     seq_len = 10
@@ -26,7 +34,7 @@ def test_feed_forward_block():
     dropout = 0.1
 
     # Create an instance of FeedForwardBlock
-    model = QuantumFeedForward(embed_len, dropout)
+    model = QuantumFeedForward(num_layers, num_wires, quantum_nn, embed_len, dropout)
 
     # Create a dummy input tensor
     dummy_input = torch.rand(batch_size, seq_len, embed_len)
@@ -45,3 +53,5 @@ def test_feed_forward_block():
     print("Test passed!")
 
     return output.shape
+
+test_feed_forward_block(num_layers=2, num_wires=6,quantum_nn=qnn_circuit,embed_len=64)
